refactor(mon-pix): move modulix components out of pods
