+++ mon-pix/app/components/module/beta-banner.hbs
