+++ mon-pix/app/components/module/details.hbs
