+++ mon-pix/app/components/module/module-element.js
