+++ mon-pix/app/components/module/objectives.hbs
