+++ mon-pix/app/components/module/passage.hbs
@@ -1,5 +1,5 @@
 {{page-title @module.title}}
-<BetaBanner />
+<Module::BetaBanner />
 
 <main class="module-passage">
   <div class="module-passage__title">
