+++ mon-pix/app/components/module/recap.hbs
@@ -1,4 +1,4 @@
-<BetaBanner />
+<Module::BetaBanner />
 
 <main class="module-recap">
   <div class="module-recap__header">
