+++ mon-pix/app/styles/app.scss
@@ -58,6 +58,7 @@ of an adaptative/mobile-first approach — refactoring is welcome here */
 @import 'components/levelup-notif';
 @import 'components/login-form';
 @import 'components/login-or-register';
+@import 'components/module/beta-banner';
 @import 'components/module/details';
 @import 'components/module/grain';
 @import 'components/module/grain/tag';
