+++ mon-pix/app/styles/components/module/_beta-banner.scss
